Fixed country name update check
before, it checked if the chars in the country name from the update were 1) in the alphabet or 2) was an underscore but that was not working so I had to embed a separate if statement in the .isalpha() check for it to work. so first it checks if its a letter and if it is not a letter if goes into a separate if statement that asks if it is an "_". if it is not an underscore, it prints to the bad updates file.

Now I need to figure out how to replace spaces and how to use all the good updates to update the actual catalogue